# core/backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Lab-Core Backend API
  version: 0.2.0
  description: |
    Lab-Core backend API の正規仕様です。
    実行単位は Job ではなく Operation を正とし、旧 `/api/jobs` と旧操作 API は含みません。
servers:
  - url: /
...
paths:
...
  /api/applications:
    get:
      tags: [Applications]
      operationId: listApplications
      summary: 削除済みを除く Application 一覧を取得
      responses:
        "200":
          description: Application 一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  applications:
                    type: array
                    items:
                      $ref: '#/components/schemas/ApplicationListItem'
                required: [applications]
    post:
      tags: [Applications]
      operationId: createApplication
      summary: Application 設定を登録
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateApplicationRequest'
      responses:
        "201":
          description: Application 登録完了
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateApplicationResponse'
        "400":
          $ref: '#/components/responses/ErrorResponse'
        "409":
          $ref: '#/components/responses/ErrorResponse'
...
  /api/applications/{applicationId}:
    get:
      tags: [Applications]
      operationId: getApplication
      summary: Application 詳細を取得
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        "200":
          description: Application 詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetailResponse'
        "404":
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags: [Applications]
      operationId: patchApplication
      summary: Application 設定を更新
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateApplicationRequest'
      responses:
        "200":
          description: 更新後の Application 詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetailResponse'
        "400":
          $ref: '#/components/responses/ErrorResponse'
        "404":
          $ref: '#/components/responses/ErrorResponse'
        "409":
          $ref: '#/components/responses/ErrorResponse'
  /api/applications/{applicationId}/deployment:
    get:
      tags: [Deployments]
      operationId: getDeployment
      summary: Deployment 設定を取得
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        "200":
          description: Deployment 設定
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDeployment'
        "404":
          $ref: '#/components/responses/ErrorResponse'
    patch:
      tags: [Deployments]
      operationId: patchDeployment
      summary: Deployment 設定を更新
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDeploymentRequest'
      responses:
        "200":
          description: 更新後の Application 詳細
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationDetailResponse'
        "400":
          $ref: '#/components/responses/ErrorResponse'
        "404":
          $ref: '#/components/responses/ErrorResponse'
        "409":
          $ref: '#/components/responses/ErrorResponse'
  /api/applications/{applicationId}/deployment/inspection:
    get:
      tags: [Deployments]
      operationId: inspectDeployment
      summary: 現在の Deployment compose を解析
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - in: query
          name: composePath
          schema:
            type: string
          required: false
          description: 省略時は現在の compose_path
      responses:
        "200":
          description: compose inspection
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: '#/components/responses/ErrorResponse'
  /api/applications/{applicationId}/operations:
    get:
      tags: [Operations]
      operationId: listOperationsByApplication
      summary: Application に紐づく Operation 一覧を取得
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
      responses:
        "200":
          description: Operation 一覧
          content:
            application/json:
              schema:
                type: object
                properties:
                  applicationId:
                    type: string
                  operations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Operation'
                required: [applicationId, operations]
        "404":
          $ref: '#/components/responses/ErrorResponse'
    post:
      tags: [Operations]
      operationId: createOperation
      summary: Application に対する Operation を作成
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOperationRequest'
      responses:
        "202":
          description: 作成された Operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationDetail'
        "400":
          $ref: '#/components/responses/ErrorResponse'
        "404":
          $ref: '#/components/responses/ErrorResponse'
        "409":
          $ref: '#/components/responses/ErrorResponse'
  /api/applications/{applicationId}/runtime-logs:
    get:
      tags: [RuntimeLogs]
      operationId: getRuntimeLogs
      summary: 起動済みコンテナの Runtime Log を取得
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
        - in: query
          name: service
          schema:
            type: string
          required: false
        - in: query
          name: tail
          schema:
            type: integer
            minimum: 20
            maximum: 1000
            default: 200
          required: false
      responses:
        "200":
          description: Runtime Log snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeLogResponse'
        "404":
          $ref: '#/components/responses/ErrorResponse'
...
components:
  parameters:
    ApplicationId:
      name: applicationId
      in: path
      required: true
      schema:
        type: string
...
  responses:
    ErrorResponse:
      description: 標準エラーレスポンス
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
              nullable: true
            requestId:
              type: string
          required: [code, message, requestId]
      required: [error]
    CreateApplicationRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        repositoryUrl:
          type: string
          format: uri
        defaultBranch:
          type: string
        composePath:
          type: string
        publicServiceName:
          type: string
        publicPort:
          type: integer
        hostname:
          type: string
        mode:
          type: string
          enum: [standard, headless]
        keepVolumesOnRebuild:
          type: boolean
        deviceRequirements:
          type: array
          items:
            type: string
        envOverrides:
          type: object
          additionalProperties:
            type: string
      required:
        - name
        - description
        - repositoryUrl
        - defaultBranch
        - composePath
        - publicServiceName
        - publicPort
        - hostname
        - mode
        - keepVolumesOnRebuild
        - deviceRequirements
        - envOverrides
    CreateApplicationResponse:
      type: object
      properties:
        applicationId:
          type: string
        deploymentId:
          type: string
        routeId:
          type: string
        message:
          type: string
      required: [applicationId, deploymentId, routeId, message]
    UpdateApplicationRequest:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        repositoryUrl:
          type: string
          format: uri
        defaultBranch:
          type: string
      additionalProperties: false
    UpdateDeploymentRequest:
      type: object
      properties:
        composePath:
          type: string
        publicServiceName:
          type: string
        publicPort:
          type: integer
        hostname:
          type: string
        keepVolumesOnRebuild:
          type: boolean
        envOverrides:
          type: object
          additionalProperties:
            type: string
      required: [composePath, publicServiceName, publicPort, hostname]
      additionalProperties: false
    CreateOperationRequest:
      type: object
      properties:
        type:
          type: string
          enum: [deploy, restart, stop, resume, rebuild, update-check, update, rollback, delete]
        parameters:
          type: object
          additionalProperties: true
      required: [type]
      additionalProperties: false
    ApplicationListItem:
      type: object
      properties:
        application_id:
          type: string
        name:
          type: string
        description:
          type: string
        repository_url:
          type: string
        default_branch:
          type: string
        current_commit:
          type: string
          nullable: true
        previous_commit:
          type: string
          nullable: true
        status:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        hostname:
          type: string
        public_service_name:
          type: string
        public_port:
          type: integer
        mode:
          type: string
        enabled:
          type: boolean
        has_update:
          type: boolean
        health:
          type: object
          nullable: true
      required:
        - application_id
        - name
        - description
        - repository_url
        - default_branch
        - status
        - created_at
        - updated_at
    ApplicationDeployment:
      type: object
      properties:
        deployment_id:
          type: string
        application_id:
          type: string
        compose_path:
          type: string
        compose_project_name:
          type: string
          nullable: true
        public_service_name:
          type: string
        public_port:
          type: integer
        hostname:
          type: string
        mode:
          type: string
        keep_volumes_on_rebuild:
          type: boolean
        device_requirements:
          type: array
          items:
            type: string
        env_overrides:
          type: object
          additionalProperties:
            type: string
        enabled:
          type: boolean
        released_at:
          type: string
          format: date-time
          nullable: true
      required:
        - deployment_id
        - application_id
        - compose_path
        - public_service_name
        - public_port
        - hostname
        - mode
        - keep_volumes_on_rebuild
        - device_requirements
        - env_overrides
        - enabled
    ApplicationDetailResponse:
      type: object
      properties:
        application:
          type: object
        deployment:
          $ref: '#/components/schemas/ApplicationDeployment'
          nullable: true
        health:
          type: object
          nullable: true
        composeInspection:
          type: object
          nullable: true
        routes:
          type: array
          items:
            type: object
        containers:
          type: array
          items:
            type: object
        updateInfo:
          type: object
          nullable: true
        events:
          type: array
          items:
            type: object
        operations:
          type: array
          items:
            $ref: '#/components/schemas/Operation'
      required: [application, deployment, health, composeInspection, routes, containers, updateInfo, events, operations]
    Operation:
      type: object
      properties:
        operationId:
          type: string
        applicationId:
          type: string
        type:
          type: string
          enum: [deploy, restart, stop, resume, rebuild, update-check, update, rollback, delete]
        status:
          type: string
          enum: [queued, running, succeeded, failed, cancelled, interrupted]
        currentStepId:
          type: string
          nullable: true
        currentStepName:
          type: string
          nullable: true
        currentStepOrder:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        parameters:
          type: object
          additionalProperties: true
        result:
          type: object
          additionalProperties: true
          nullable: true
        error:
          type: object
          nullable: true
        retryOfOperationId:
          type: string
          nullable: true
        logsAvailable:
          type: boolean
        canCancel:
          type: boolean
        canRetry:
          type: boolean
        links:
          type: object
          properties:
            self:
              type: string
            logs:
              type: string
            logStream:
              type: string
          required: [self, logs, logStream]
      required:
        - operationId
        - applicationId
        - type
        - status
        - createdAt
        - updatedAt
        - parameters
        - logsAvailable
        - canCancel
        - canRetry
        - links
    OperationStep:
      type: object
      properties:
        stepId:
          type: string
        operationId:
          type: string
        stepOrder:
          type: integer
        name:
          type: string
        status:
          type: string
          enum: [pending, running, succeeded, failed, skipped]
        startedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
          nullable: true
        message:
          type: string
          nullable: true
        errorCode:
          type: string
          nullable: true
        details:
          type: object
          additionalProperties: true
          nullable: true
      required: [stepId, operationId, stepOrder, name, status, updatedAt]
    OperationDetail:
      allOf:
        - $ref: '#/components/schemas/Operation'
        - type: object
          properties:
            steps:
              type: array
              items:
                $ref: '#/components/schemas/OperationStep'
          required: [steps]
...
    RuntimeLogResponse:
      type: object
      properties:
        applicationId:
          type: string
        applicationName:
          type: string
        service:
          type: string
          nullable: true
        services:
          type: array
          items:
            type: string
        tail:
          type: integer
        lines:
          type: array
          items:
            type: string
        fetchedAt:
          type: string
          format: date-time
        executionMode:
          type: string
          enum: [dry-run, execute]
      required: [applicationId, applicationName, service, tail, lines, fetchedAt, executionMode]